Add f64 tests and split off test from lib.rs into a test file
